Fix warnings

+++ src/num/gpuops.c
@@ -97,7 +97,7 @@ int cuda_get_device(void)
 
 static int num_cuda_reserved_devices(void)
 {
-	assert(bitcount(reserved_gpus) == n_reserved_gpus);
+	assert((int)bitcount(reserved_gpus) == n_reserved_gpus);
 	return n_reserved_gpus;
 }
